--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,4 @@
 *.class
 code
 .idea/
+*.beam