--- a/clean.sh
+++ b/clean.sh
@@ -4,3 +4,4 @@ rm jvm/code.class
 rm -r rust/target
 rm -rf kotlin/code.jar
 rm dart/code
+rm Elixir.Bench.beam