--- a/fibonacci/elixir/bench.exs
+++ b/fibonacci/elixir/bench.exs
@@ -1,0 +1,1 @@
+Bench.run()