--- a/loops/elixir/bench.exs
+++ b/loops/elixir/bench.exs
@@ -1,0 +1,1 @@
+Bench.run()