This is a recursion based Elixir implementation. Initially implemented as streams, it was slow, then using reduce, then came to a solution based on recursive function calls. Implementation using for loops doesn't really make sense since the data in elixir is immutable

fibonacci:
```
time elixir elixir/bench.exs 40
165580140

real    0m1.982s
user    0m2.660s
sys     0m0.305s
```
loops:
```
time elixir elixir/bench.exs 40
1952939

real    0m3.490s
user    0m4.178s
sys     0m0.251s
```