@michelson 2024-11-26
https://gist.github.com/michelson/fcf05b0f7a7fe1aba95b98a5efff43ee#file-bench-ex
This elixir version takes an Execution Time: of 1.025 seconds