@josevalim 2024-11-26
Nice work @klirix, some notes:

1. You don't need to compile the Elixir file. You can just run the script directly. Here is what I wrote for fib:

```
defmodule Math do
  def fib(0), do: 0
  def fib(1), do: 1
  def fib(n), do: fib(n - 1) + fib(n - 2)

  def loop(0, res), do: IO.puts(res)
  def loop(i, res), do: loop(i - 1, res + fib(i))
end

[num] = System.argv()
Math.loop(String.to_integer(num) - 1, 0)
```

2. We should not use tuples for loops in Elixir. Because the initial benchmark assumes mutability, it is unclear if that's a requirement, so we have two options:

a. Build a list as we build the loops
b. Use the `:counters` module (https://www.erlang.org/doc/apps/erts/counters.html), which is literally a mutable array

Perhaps option `b` will keep the spirit of the benchmark.