@klirix 2024-11-26
Thanks @josevalim
> You don't need to compile the Elixir file. You can just run the script directly
Wouldn't precompilation save time on initialization times and allow for optimizations?

> 2. We should not use tuples for loops in Elixir. Because the initial benchmark assumes mutability, it is unclear if that's a requirement, so we have two options:
I somehow got slightly better times using the Tuples instead of the Lists in the interation itself, but also I used it for faster random element access in the end of the benchmark itself. My initial implementation was using Lists but I thought that preallocation would make sense here, even though it goes against idiomatic solutions

> b. Use the `:counters` module (https://www.erlang.org/doc/apps/erts/counters.html), which is literally a mutable array
> Perhaps option `b` will keep the spirit of the benchmark.
I was thinking the same about the spirit of menchmark maybe using counters module would be a better solution