@josevalim 2024-11-26
Scripts are also compiled before running, so runtime performance is the same. And the file is small enough that the time to compile it is negligible.

counters should be more efficient than tuples too :)