@DeedleFake 2024-11-26
@josevalim

a. I tried doing that and it wound up _slower_ than the tuple-based version. I've been trying to figure out why. See this comment: https://github.com/bddicken/languages/pull/36#discussion_r1858817151

b. Ooh. I tried using `:ets.update_counter/4` and that _kind of_ worked, but it was both kind of ugly code and quite slow. `:counters` is good to know about. Unfortunately, I tried building a version that used that it was a lot slower than just building a list:

```elixir
defmodule New do
  def run() do
    [input | _] = System.argv()
    input = input |> String.to_integer()
    rand = :rand.uniform(10_000)
    result = outer(rand, input)
    IO.puts(result)
  end

  defp outer(i \\ 1, rand, input),
    do: outer(i, rand, input, :counters.new(10_000, [:write_concurrency]))

  defp outer(10_001, rand, _input, counters), do: :counters.get(counters, rand + 1)

  defp outer(i, rand, input, counters) do
    :counters.add(counters, i, inner(input) + rand)
    outer(i + 1, rand, input, counters)
  end

  defp inner(i2 \\ 0, input, result \\ 0)
  defp inner(100_000, _input, result), do: result

  defp inner(i2, input, result) do
    result = result + rem(i2, input)
    inner(i2 + 1, input, result)
  end
end
```

Benchee results:

```
Benchmarking new ...
1954561
1950948
1955270
1957398
Benchmarking old ...
1950065
1950118
1955649
1957186
Calculating statistics...
Formatting results...

Name           ips        average  deviation         median         99th %
old           0.49         2.03 s     ±0.10%         2.03 s         2.03 s
new           0.42         2.41 s     ±0.81%         2.40 s         2.43 s

Comparison:
old           0.49
new           0.42 - 1.19x slower +0.38 s
```