@josevalim 2024-11-26
@DeedleFake the list one may be slower because of the range? You can try this:

```
@list Enum.to_list(1..9999)

def list(...) do
  for _ <- @list do

  end
end
```

There is a chance the cost is in the protocol, which is not consolidated if you are not inside the project.