@klirix 2024-11-26
@michelson 

My intuition coming from other languages was telling me that dynamic allocations could be a culprit here. I know there are tons of optimizations for tuples in the vm for inserting and fetching the tuple elements and it's not actually allocating the tuple 10k times. But the list is actually allocating each node on iteration, so that makes for 10k miniscule allocation operations.

In all fairness 10k is a pretty small scale anyway to see the effects