@DeedleFake 2024-11-26
@klirix

You should actually be using `put_elem/3`, not `Tuple.insert_at/3`. You have an existing tuple with 10,000 zeros. It should be changing them, not inserting new ones.