@DeedleFake 2024-11-26
It seems that I've misunderstood the way that pattern matching and guard clauses are optimized. It turns out that this order is quite a bit faster, making it the fastest one that I've tested so far:

```elixir
  def run() do
    [input] = System.argv()
    input = input |> String.to_integer()
    rand = :rand.uniform(9_999) - 1
    result = outer(rand, input)
    IO.puts(result)
  end

  defp outer(i \\ 0, rand, input, results \\ [])

  defp outer(i, rand, input, results) when i < 10_000 do
    result = inner(input) + rand
    outer(i + 1, rand, input, [result | results])
  end

  defp outer(_i, rand, _input, results), do: results |> Enum.at(10_000 - rand - 1)

  defp inner(i2 \\ 0, input, result \\ 0)

  defp inner(i2, input, result) when i2 < 100_000 do
    result = result + rem(i2, input)
    inner(i2 + 1, input, result)
  end

  defp inner(_i2, _input, result), do: result
```